Added possibility to specify encoding for scripts and input files as well as output encoding.
and fixed unit tests for new encoding capabilities.
